fixed all math to use BigDecimal instead of doubles

+++ module-1/capstone/java/src/main/java/com/techelevator/Candy.java
@@ -1,8 +1,10 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Candy extends Item {
 	// CTOR
-	public Candy(String name, String slot, double price) {
+	public Candy(String name, String slot, BigDecimal price) {
 		super(name, slot, price);
 		this.type = "Candy";
 		this.sound = "Munch Munch, Yum!";
